#61 some dependencies moved from core to repeat
