+++ pom.xml
@@ -17,7 +17,6 @@
 
     <properties>
         <log4j2>2.10.0</log4j2>
-        <quartz>2.3.0</quartz>
         <bot-api>3.5</bot-api>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
@@ -34,24 +33,9 @@
             <version>${bot-api}</version>
         </dependency>
         <dependency>
-            <groupId>org.telegram</groupId>
-            <artifactId>telegrambots-abilities</artifactId>
-            <version>3.6</version>
-        </dependency>
-        <dependency>
             <groupId>postgresql</groupId>
             <artifactId>postgresql</artifactId>
             <version>9.1-901-1.jdbc4</version>
-        </dependency>
-        <dependency>
-            <groupId>org.quartz-scheduler</groupId>
-            <artifactId>quartz</artifactId>
-            <version>${quartz}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.quartz-scheduler</groupId>
-            <artifactId>quartz-jobs</artifactId>
-            <version>${quartz}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.commons</groupId>
@@ -134,7 +118,6 @@
             <version>1.4.3</version>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
 
     <build>
