+++ repeat-bot/src/main/java/ru/ustits/colleague/repeats/tools/CronFields.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.cron;
+package ru.ustits.colleague.repeats.tools;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
